# Factor out and improve the vectorization of RegexInterpreter.FindFirstChar