--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Regex.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Regex.cs
@@ -65,12 +65,12 @@ namespace System.Text.RegularExpressions
             // Call Init directly rather than delegating to a Regex ctor that takes
             // options to enable linking / tree shaking to remove the Regex compiler
             // and NonBacktracking implementation if it's not used.
-            Init(pattern, RegexOptions.None, s_defaultMatchTimeout, culture);
+            Init(pattern, RegexOptions.None, s_defaultMatchTimeout, culture ?? CultureInfo.CurrentCulture);
         }
 
         internal Regex(string pattern, RegexOptions options, TimeSpan matchTimeout, CultureInfo? culture)
         {
-            culture ??= GetTargetCulture(options);
+            culture ??= RegexParser.GetTargetCulture(options);
             Init(pattern, options, matchTimeout, culture);
 
             if ((options & RegexOptions.NonBacktracking) != 0)
@@ -87,10 +87,6 @@ namespace System.Text.RegularExpressions
             }
         }
 
-        /// <summary>Gets the culture to use based on the specified options.</summary>
-        private static CultureInfo GetTargetCulture(RegexOptions options) =>
-            (options & RegexOptions.CultureInvariant) != 0 ? CultureInfo.InvariantCulture : CultureInfo.CurrentCulture;
-
         /// <summary>Initializes the instance.</summary>
         /// <remarks>
         /// This is separated out of the constructor so that an app only using 'new Regex(pattern)'
@@ -98,7 +94,7 @@ namespace System.Text.RegularExpressions
         /// compiler, such that a tree shaker / linker can trim it away if it's not otherwise used.
         /// </remarks>
         [MemberNotNull(nameof(_code))]
-        private void Init(string pattern, RegexOptions options, TimeSpan matchTimeout, CultureInfo? culture)
+        private void Init(string pattern, RegexOptions options, TimeSpan matchTimeout, CultureInfo culture)
         {
             ValidatePattern(pattern);
             ValidateOptions(options);
@@ -107,7 +103,6 @@ namespace System.Text.RegularExpressions
             this.pattern = pattern;
             internalMatchTimeout = matchTimeout;
             roptions = options;
-            culture ??= GetTargetCulture(options);
 
 #if DEBUG
             if (IsDebug)
@@ -121,7 +116,7 @@ namespace System.Text.RegularExpressions
 
             // Generate the RegexCode from the node tree.  This is required for interpreting,
             // and is used as input into RegexOptions.Compiled and RegexOptions.NonBacktracking.
-            _code = RegexWriter.Write(tree);
+            _code = RegexWriter.Write(tree, culture);
 
             if ((options & RegexOptions.NonBacktracking) != 0)
             {
@@ -434,7 +429,7 @@ namespace System.Text.RegularExpressions
         /// <summary>Creates a new runner instance.</summary>
         private RegexRunner CreateRunner() =>
             factory?.CreateInstance() ??
-            new RegexInterpreter(_code!, GetTargetCulture(roptions));
+            new RegexInterpreter(_code!, RegexParser.GetTargetCulture(roptions));
 
         /// <summary>True if the <see cref="RegexOptions.Compiled"/> option was set.</summary>
         protected bool UseOptionC() => (roptions & RegexOptions.Compiled) != 0;