--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/Algebras/BDD.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/Algebras/BDD.cs
@@ -240,6 +240,7 @@ namespace System.Text.RegularExpressions.Symbolic
         /// Serializer uses more compacted representations when fewer bits are needed, which is reflected in the first
         /// two numbers of the return value. MTBDD terminals are represented by negated numbers as -id.
         /// </summary>
+        [ExcludeFromCodeCoverage]
         public long[] Serialize()
         {
             if (IsEmpty)