… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexBoyerMoore.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexFindOptimizations.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexInterpreter.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexNode.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexParser.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexPrefixAnalyzer.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexWriter.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexInfo.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexMatcher.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexNode.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexRunnerFactory.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexSampler.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/Unicode/GeneratorHelper.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/Unicode/IgnoreCaseRelationGenerator.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/Unicode/UnicodeCategoryRangesGenerator.cs
… src/libraries/System.Text.RegularExpressions/src/System/Threading/StackHelper.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Groups.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Match.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Tests.Common.cs
… src/libraries/System.Text.RegularExpressions/tests/RegexCultureTests.cs
… src/libraries/System.Text.RegularExpressions/tests/RegexExperiment.cs
… src/libraries/System.Text.RegularExpressions/tests/RegexReductionTests.cs