This change started with the "simple" goal of factoring out the FindFirstChar logic from RegexInterpreter and consuming it in SymbolicRegexMatcher.  The existing engines use FindFirstChar to quickly skip ahead to the next location that might possibly match, at which point they fall back to analyzing the whole pattern at that location.  SymbolicRegexMatcher (used by RegexOptions.NonBacktracking) had its own implementation for this, which it used any time it entered a start state.  This required non-trivial additional code to maintain, and there's no good reason it should be separate from the other engines.

However, what started out as a simple change grew due to regressions that resulted from differences in the implementations.  In particular, SymbolicRegexMatcher already works off of precomputed equivalence tables for casing, which gives it very different characteristics in this regard from the existing engines.  For example, SymbolicRegexMatcher's existing "skip ahead to the next possible match start location" logic already evaluated all the characters that could possibly start a match, which included variations of the same character when using IgnoreCase, but the existing RegexInterpreter logic didn't.  That discrepancy then results in a significant IgnoreCase regression for NonBacktracking due to losing the ability to use a vectorized search for the next starting location.  We already plan to shift the existing engines over to a plan where all of these equivalences are computed at construction time rather than using ToLower at both construction time and match time, so this PR takes some steps in that direction, doing so for most of ASCII.  This has added some temporary cruft, which we'll be able to delete once we fully shift the implementations over (which we should do in the near future).

Another difference was SymbolicRegexMatcher was enabling use of IndexOfAny for up to 5 characters, whereas RegexOptions.Compiled was only doing up to 3 characters, and RegexInterpreter wasn't doing for any number.  The PR now uses 5 everywhere.

However, the more characters involved, the more overhead there is to IndexOfAny, and for some inputs, the higher the chances are that IndexOfAny will find a match sooner, which means its overhead compounds more.  To help with that, we now not only compute the possible characters that might match at the beginning of the pattern, but also characters that might match at a fixed offset from the beginning of the pattern (e.g. in \d{3}-\d{2}-\d{4}, it will find the '-' at offset 3 and be able to vectorize a search for that and then back off by the relevant distance.  That then also means we might end up with multiple sets to choose to search for, and this PR borrows an idea from Rust, which is to use some rough frequency analysis to determine which set should be targeted.  It's not perfect, and we can update the texts use to seed the analysis (right now I based it primarily on *.cs files in dotnet/runtime and some Project Gutenberg texts), but it's good enough for these purposes for now.

We'd previously switched to using IndexOf for a case-sensitive prefix string, but still were using Boyer-Moore for case-insensitive.  Now that we're able to also vectorize a search for case-insensitive values (right now just ASCII letter, but that'll be fixed soon), we can just get rid of Boyer-Moore entirely.  This saves all the costs to do with constructing the Boyer-Moore tables and also avoids having to generate the Boyer-Moore implementations in RegexOptions.Compiled and the source generator.

The casing change also defeated some other optimizations already present.  For example, in .NET 5 we added an optimization whereby an alternation like `abcef|abcgh` would be transformed into `abc(?:ef|gh)`, and that would apply whether case-sensitive or case-insensitive.  But by transforming the expression at construction now for case-insensitive into `[Aa][Bb][Cc][Ee][Ff]|[Aa][Bb][Cc][Gg][Hh]`, that optimization was defeated.  I've added a new optimization pass for alternations that will detect common prefixes even if they're sets.

The casing change also revealed some cosmetic issues.  As part of the change, when we encounter a "multi" (a multi-character string in the pattern), we convert that single case-insensitive RegexNode to instead be one case-sensitive RegexNode per character, with a set for all the equivalent characters that can match.  This then defeats some of the nice formatting we had for multis in the source generator, so as part of this change, the source generator has been augmented to output nicer code for concatenations.  And because sets like [Ee] are now way more common (since e.g. a case-insensitive 'e' will be transformed into such a set), we also special-case that in both the source generator and RegexOptions.Compiled, to spit out the equivalent of `(c | 0x20) == 'e'` rather than `(c == 'E'| c == 'e')`.

Along the way, I cleaned up a few things as well, such as passing around a CultureInfo more rather than repeatedly calling CultureInfo.CurrentCulture, using CollectionsMarshal.GetValueRefOrAddDefault on a hot path to do with interning strings in a lookup table, tweaking SymbolicRegexRunnerFactory's Runner to itself be generic to avoid an extra layer of virtual dispatch per operation, and cleaning up code / comments in SymbolicRegexMatcher along the way.

For the most part the purpose of the change wasn't to improve perf, and in fact I was willing to accept some regressions in the name of consolidation.  There are a few regressions here, mostly small, and mostly for cases where we're simply paying an overhead for vectorization, e.g. where the current location is fine to match, or where the target character being searched for is very frequent.  Overall, though, there are some substantial improvements.

<details>
<summary>Benchmarks</summary>

|                                  Type |                 Method |         Toolchain |                              Pattern |                     Options |                Mean | Ratio |       Allocated |
|-------------------------------------- |----------------------- |------------------ |------------------------------------- |---------------------------- |--------------------:|------:|----------------:|
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? |                  IgnoreCase | 5,919,020,750.00 ns |  1.00 |   229,137,752 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? |                  IgnoreCase | 2,655,622,120.00 ns |  0.45 |   229,135,392 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? |        IgnoreCase, Compiled | 5,980,794,825.00 ns |  1.00 |   229,135,392 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? |        IgnoreCase, Compiled | 1,709,189,240.00 ns |  0.29 |   229,135,392 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? | IgnoreCase, NonBacktracking | 6,274,067,325.00 ns |  1.00 |   229,135,392 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? | IgnoreCase, NonBacktracking | 2,792,333,450.00 ns |  0.45 |   229,135,392 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? |                        None | 3,769,231,500.00 ns |  1.00 |   114,408,208 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? |                        None | 3,714,378,675.00 ns |  0.99 |   114,408,208 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? |                    Compiled | 3,706,293,650.00 ns |  1.00 |   150,863,280 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? |                    Compiled | 3,710,450,780.00 ns |  1.00 |   114,408,160 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
|        Perf_Regex_Industry_SliceSlice |                  Count | \main\corerun.exe |                                    ? |             NonBacktracking | 4,208,438,720.00 ns |  1.00 |   114,408,208 B |
|        Perf_Regex_Industry_SliceSlice |                  Count |   \pr\corerun.exe |                                    ? |             NonBacktracking | 3,745,114,350.00 ns |  0.89 |   114,408,208 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                           (?i)Holmes |                        None |       789,132.09 ns |  1.00 |        97,138 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                           (?i)Holmes |                        None |       688,469.87 ns |  0.87 |        97,138 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                           (?i)Holmes |                    Compiled |       808,262.69 ns |  1.00 |        97,138 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                           (?i)Holmes |                    Compiled |       417,193.55 ns |  0.52 |        97,137 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                           (?i)Holmes |             NonBacktracking |       725,918.41 ns |  1.00 |        97,138 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                           (?i)Holmes |             NonBacktracking |       685,504.44 ns |  0.94 |        97,140 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |                        None |     6,870,657.12 ns |  1.00 |       145,019 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |                        None |     4,824,624.36 ns |  0.70 |       144,990 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |                    Compiled |     2,258,139.88 ns |  1.00 |       144,984 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |                    Compiled |       825,089.83 ns |  0.37 |       144,979 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |             NonBacktracking |     1,890,083.74 ns |  1.00 |       144,981 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |             (?i)Sher[a-z]+\|Hol[a-z]+ |             NonBacktracking |     1,388,542.11 ns |  0.73 |       144,984 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                         (?i)Sherlock |                        None |       471,244.46 ns |  1.00 |        21,217 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                         (?i)Sherlock |                        None |       150,869.37 ns |  0.32 |        21,216 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                         (?i)Sherlock |                    Compiled |       523,646.53 ns |  1.00 |        21,218 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                         (?i)Sherlock |                    Compiled |        99,295.06 ns |  0.19 |        21,216 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                         (?i)Sherlock |             NonBacktracking |       454,725.55 ns |  1.00 |        21,217 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                         (?i)Sherlock |             NonBacktracking |       144,565.38 ns |  0.32 |        21,217 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                  (?i)Sherlock Holmes |                        None |       400,282.70 ns |  1.00 |        19,970 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                  (?i)Sherlock Holmes |                        None |       148,763.96 ns |  0.37 |        19,968 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                  (?i)Sherlock Holmes |                    Compiled |       320,484.07 ns |  1.00 |        19,969 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                  (?i)Sherlock Holmes |                    Compiled |        97,347.57 ns |  0.30 |        19,968 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                  (?i)Sherlock Holmes |             NonBacktracking |       374,497.60 ns |  1.00 |        19,969 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                  (?i)Sherlock Holmes |             NonBacktracking |       148,935.01 ns |  0.40 |        19,968 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |                        None |     9,336,569.80 ns |  1.00 |       135,227 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |                        None |     8,944,893.54 ns |  0.96 |       135,258 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |                    Compiled |     2,388,718.93 ns |  1.00 |       135,209 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |                    Compiled |     1,685,141.16 ns |  0.71 |       135,205 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |             NonBacktracking |     2,627,173.71 ns |  1.00 |       135,213 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |           (?i)Sherlock\|Holmes\|Watson |             NonBacktracking |     3,059,709.81 ns |  1.16 |       135,209 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe | (?i)Sherlock\|(...)er\|John\|Baker [49] |                        None |    25,380,686.43 ns |  1.00 |       156,696 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe | (?i)Sherlock\|(...)er\|John\|Baker [49] |                        None |    30,027,019.05 ns |  1.18 |       156,775 B |
|                                       |                        |                   |                                      |                             |                     |       |                 |
…