@stephentoub 2021-11-12
@olsaarik, @veanes, please take a look at the NonBacktracking changes when you get a moment.

Also, this specific improvement stands out:

|                                  Type |                 Method |         Toolchain |                              Pattern |                     Options |                Mean | Ratio |       Allocated |
|-------------------------------------- |----------------------- |------------------ |------------------------------------- |---------------------------- |--------------------:|------:|----------------:|
| Perf_Regex_Industry_RustLang_Sherlock |                  Count | \main\corerun.exe |                     [a-q][^u-z]{13}x |             NonBacktracking | 1,913,033,821.43 ns | 1.000 | 2,250,206,424 B |
| Perf_Regex_Industry_RustLang_Sherlock |                  Count |   \pr\corerun.exe |                     [a-q][^u-z]{13}x |             NonBacktracking |        99,301.51 ns | 0.000 |        29,536 B |

and while it's great that it improved so much, that improvement comes just from skipping ahead, and these numbers suggest that if it wasn't able to skip ahead like that, there's some kind of real perf issue lurking.