@danmoseley 2021-11-12
Awesome.

Looking at the perf results - for the scenario with the value 136.19 ns, why is Compiled 10,000 times faster than the others?