@stephentoub 2021-11-12
> Looking at the perf results - for the scenario with the value 136.19 ns, why is Compiled 10,000 times faster than the others?

One of the advantages of the extra analysis that's done in compilation. The pattern here is `(?s).*`.  That's single-line mode doing a `.*`, which means `match all of anything`.  The compiled code recognizes that and simply jumps to the end.
https://github.com/dotnet/runtime/blob/5fa6dd364982be4ffd83358adbf130d88049c72a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCompiler.cs#L2961-L2972