@danmoseley 2021-11-12
>  The compiled code recognizes that and simply jumps to the end.

Presumably the interpreter/backtracking could easily jump to the end as well?
Actually, I wonder whether the regex writer stage could transform this to something, eg., `(?s)$` that would have the same effect.