@stephentoub 2021-11-12
> Presumably the interpreter/backtracking could easily jump to the end as well?

Sure, if it added a check at execution time.

> I wonder whether the regex writer stage could transform this to something, eg., (?s)$ that would have the same effect.

Is this really a pattern we care to optimize?  The optimization in the compiler is there to better handle the case where this is in the middle of the pattern, and it simply carries over to this corner-case.