@stephentoub 2021-11-12
> Some of the rows in your table seem to be broken because | was not escaped

Ugh.  I manually escaped a bunch of things, but I'll go back and see what I missed.

> maybe we should open Benchmark.NET bug?

https://github.com/dotnet/BenchmarkDotNet/issues/1839