@stephentoub 2021-11-12
> eg., Baker [49] rows.

Hmm, I'm not seeing any that are broken.  I fixed a bunch yesterday within minutes after opening the PR and seeing the output.  Have you refreshed?  Or, if they're still there, which lines exactly?