@danmoseley 2021-11-12
Ah, yes I just hadn't refreshed. Still it's good you opened the  bug.