@stephentoub 2021-11-16
> Also, this specific improvement stands out

Profiling it, this particular issue is due to the Antimirov (NFA) mode.  It's currently really slow, and this particular pattern ends up tripping over the 10K DFA state limit (it ends up with something like 15K states if I let it) that puts it into that mode; the monstrous perf gap goes away if I lift the state limit. https://github.com/dotnet/runtime/issues/60918 tracks at least one aspect of improving its performance.
cc: @olsaarik