@stephentoub 2021-11-17
@danmoseley, @joperezr, are you comfortable with this being merged?  It touches a lot of code and it'd be nice to get it in to unblock other things in the same areas, e.g. I imagine this FindFirstChar logic might be one of the first places Jose wants to touch when starting to use span more throughout as part of https://github.com/dotnet/runtime/issues/59629.