@DrewScoggins 2021-11-19
Improvements: https://github.com/dotnet/perf-autofiling-issues/issues/2424