@AndyAyersMS 2021-11-23
Improvements: https://github.com/dotnet/perf-autofiling-issues/issues/2530